backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12
groq==0.11.0
chromadb==0.5.23
sentence-transformers==3.3.1
pdfplumber==0.11.4
python-docx==1.1.2
pydantic==2.10.3
